refactor(draug): A.4a — move remaining mcp_handler agent files
Continues the file-by-file migration. Compositor still builds clean
through thin re-export shims; existing call sites resolve unchanged.

Moved to draug-daemon:
* knowledge_hunt.rs (1072 LoC) — Phase 7 Wikipedia/source fetch +
  Phase 17 refactor-step LLM prompts. Dep: fbp-rs (added to
  draug-daemon's Cargo.toml).
* agent_planner.rs (555 LoC) — Plan-and-Solve / Phase 15 LLM driver.
* refactor_loop.rs (432 LoC) — Phase 17 task selection + retry
  policy + REFACTOR_FIXTURES seed.
* draug_async.rs (979 LoC) — non-blocking AsyncOp state machine
  driving TCP for LLM/PATCH/cargo_check via tcp_connect_async.

Each file's `compositor::draug::*` import became `crate::draug::*`
(same crate now), `super::knowledge_hunt::*` etc. became
`crate::knowledge_hunt::*` (mcp_handler parent gone in the daemon
crate's flat layout), and the `extern crate alloc;` line was dropped
since lib.rs already declares it.

`pub(super)` and `pub(crate)` items inside the moved files were
promoted to plain `pub` so compositor's
`pub use draug_daemon::knowledge_hunt::*;` glob shim re-exports
them. Same change applied to the previously-moved
`draug.rs` / `task_store.rs` / `refactor_types.rs`.

Compositor mcp_handler files now one-line shims:
    pub use draug_daemon::knowledge_hunt::*;
    pub use draug_daemon::agent_planner::*;
    pub use draug_daemon::refactor_loop.rs::*;
    pub use draug_daemon::draug_async::*;

Verified: cargo check passes for libfolk, draug-daemon, compositor
and `cargo check --bin compositor`. No new errors. Warning count
shifted some from compositor to draug-daemon as expected
(dead-code warnings now reach the lib boundary).

Files NOT moved (compositor UI coupling, deferred to A.4b after A.5):
* agent_logic.rs (427 LoC) — owns WasmState/WindowManager refs
* autodream.rs (740 LoC) — owns WasmState cache eviction
* token_stream.rs (411 LoC) — owns TokenRing + think overlay UI
* mcp_handler/mod.rs (116 LoC) — orchestration entry points
